Add files via upload
One of the leading news channels CNBE who wants to analyze recent elections. This survey was conducted on 1525 voters with 9 variables. Build a model, to predict which party a voter will vote for on the basis of the given information, to create an exit poll that will help in predicting the overall win and seats covered by a particular party.
 EDA / Data analysis / Text analytics / machine learning /